feat: add verify to templates (#30)

+++ src/syncfusion/index.ts
@@ -89,6 +89,10 @@ export class IaraSyncfusionAdapter
   }
 
   insertInference(inference: IaraSpeechRecognitionDetail): void {
+    if (inference.richTranscriptModifiers) {
+      this.insertTemplate(inference.richTranscript);
+      return;
+    }
     if (inference.isFirst) {
       if (this._selectionManager.selection.text.length)
         this._editor.documentEditor.editor.delete();
